Phase 4 Wave A: move the code editor into surfaces/studio/
First batch of Creator Studio's core/+surfaces/ scaffolding (admin was
#1070, party #1072, review #1073, catalog #1077) — the code-editor
sub-cluster: EditorPanel/EditorSurface, the whole CodeSurface family
(CodeMirrorEditor, the file explorer, upload/archive/drop handling, the
in-browser TS language service and its worker, WebMCP tools), and their
small leaf utilities (diffLines, codeTokens, editorParamsScrub,
editorJsonLiveDiff, NumberScrubber, codeActionsMatch/CodeActionsMenu).

studioApi.ts, editorBridge.ts, editorContentTools.ts,
editorControllerProtocol.ts, LayeredEditorSurface.tsx, PathPainter.tsx,
roundSealed.ts, studioStatusPoll.ts, and StudioCreatorAgentKeyPanel.tsx
stay in place: each is imported well beyond this batch — by the Remix
feature (RemixPainter.tsx, RemixPanel.tsx), by party (mp/protocol.ts), by
the account-settings modal, or by Studio widgets landing in a later
batch — the same call as leaving catalog.ts/mp/protocol.ts/proposalsApi.ts
in place for the earlier surface moves.

CreatorStudioView.tsx itself stays at the root for now (it's the biggest
file in this cluster and the natural final batch, once everything it
renders has already moved); its three moved imports (CodeActionsMenu,
CodeSurface, EditorPanel) are repointed at the new paths.

## Verification

- tsc --noEmit clean
- Full apps/web suite: 215 test files / 1817 tests, all pass
- Repo npm run lint: 0 errors
- Production build succeeds; tsWorker and CreatorStudioView remain their
  own lazy chunks

## Scope note

Remaining Studio sub-batches (widget panels, connect/onboarding, and
finally CreatorStudioView.tsx itself), core/ itself, poller consolidation,
and the styles.css split land in later PRs.
