+++ apps/web/src/CreatorStudioView.tsx
@@ -11,9 +11,9 @@ import { formatRelativeTime } from './relativeTime.js';
 import { playPath, studioPath, type StudioTab } from './router.js';
 import { abandonSubmission, deleteGame, handoffToPlatform } from './submissionApi.js';
 import { StudioShotToasts } from './StudioShotToasts.js';
-import { type CodeActionsMode } from './CodeActionsMenu.js';
-import { CodeSurface } from './CodeSurface.js';
-import { EditorPanel } from './EditorPanel.js';
+import { type CodeActionsMode } from './surfaces/studio/CodeActionsMenu.js';
+import { CodeSurface } from './surfaces/studio/CodeSurface.js';
+import { EditorPanel } from './surfaces/studio/EditorPanel.js';
 import { StudioStage, type StagePosture, type StageStatus } from './StudioStage.js';
 import { StudioStrip } from './StudioStrip.js';
 import { usePlayChromeIdle } from './usePlayChromeIdle.js';
