+++ apps/web/src/surfaces/studio/CodeSurface.test.tsx
@@ -6,8 +6,8 @@ import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import { CodeSurface } from './CodeSurface.js';
 import { getCodeSurfaceSessionState, resetCodeSurfaceSessionState } from './codeSurfaceSessionState.js';
 import * as codeSurfaceApi from './codeSurfaceApi.js';
-import i18n from './i18n/index.js';
-import type { EditorContentDoc, GameEditorState } from './studioApi.js';
+import i18n from '../../i18n/index.js';
+import type { EditorContentDoc, GameEditorState } from '../../studioApi.js';
 
 const codeMirrorMock = vi.hoisted(() => ({
   current: null as {
@@ -52,8 +52,8 @@ vi.mock('./codeSurfaceApi.js', async () => {
   };
 });
 
-vi.mock('./studioApi.js', async () => {
-  const actual = await vi.importActual<typeof import('./studioApi.js')>('./studioApi.js');
+vi.mock('../../studioApi.js', async () => {
+  const actual = await vi.importActual<typeof import('../../studioApi.js')>('../../studioApi.js');
   return { ...actual, fetchGameEditor: vi.fn() };
 });
 
@@ -63,15 +63,15 @@ vi.mock('./webmcp.js', async () => {
 });
 
 // Stub the agent-mode modal's key panel fetch.
-vi.mock('./connectApi.js', async () => {
-  const actual = await vi.importActual<typeof import('./connectApi.js')>('./connectApi.js');
+vi.mock('../../connectApi.js', async () => {
+  const actual = await vi.importActual<typeof import('../../connectApi.js')>('../../connectApi.js');
   return { ...actual, getCreatorAgentKey: vi.fn() };
 });
 
 const mocked = vi.mocked(codeSurfaceApi);
-const mockedStudioApi = vi.mocked(await import('./studioApi.js'));
+const mockedStudioApi = vi.mocked(await import('../../studioApi.js'));
 const mockedWebmcp = vi.mocked(await import('./webmcp.js'));
-const mockedConnectApi = vi.mocked(await import('./connectApi.js'));
+const mockedConnectApi = vi.mocked(await import('../../connectApi.js'));
 
 async function flush() {
   await Promise.resolve();
