+++ apps/web/src/surfaces/studio/NumberScrubber.test.tsx
