+++ apps/web/src/surfaces/studio/NumberScrubber.tsx
