+++ apps/web/src/surfaces/studio/codeMirrorColors.test.ts
