+++ apps/web/src/surfaces/studio/codeMirrorColors.ts
