+++ apps/web/src/surfaces/studio/codeMirrorSearchPanel.test.ts
