+++ apps/web/src/surfaces/studio/codeMirrorSearchPanel.ts
