+++ apps/web/src/surfaces/studio/codeSurfacePaths.test.ts
