+++ eslint-rules/comment-prose-baseline.json
@@ -385,9 +385,6 @@
     "apps/web/src/ClaimHandleModal.tsx": 67,
     "apps/web/src/ClosedBetaSplash.tsx": 146,
     "apps/web/src/CodeSurface.test.tsx": 114,
-    "apps/web/src/CodeSurface.tsx": 630,
-    "apps/web/src/codeSurfaceApi.ts": 82,
-    "apps/web/src/codeTokens.ts": 177,
     "apps/web/src/connectApi.ts": 231,
     "apps/web/src/connectCollapse.ts": 26,
     "apps/web/src/ContactPage.tsx": 64,
@@ -408,9 +405,7 @@
     "apps/web/src/editorBridge.ts": 228,
     "apps/web/src/EditorCollections.test.tsx": 13,
     "apps/web/src/editorContentTools.ts": 147,
-    "apps/web/src/editorJsonLiveDiff.ts": 92,
     "apps/web/src/EditorPanel.test.tsx": 16,
-    "apps/web/src/EditorPanel.tsx": 644,
     "apps/web/src/EditProfileModal.tsx": 56,
     "apps/web/src/GameDetailPage.test.tsx": 15,
     "apps/web/src/GameDetailPage.tsx": 36,
@@ -587,6 +582,15 @@
     "apps/web/src/surfaces/review/ProposalReviewCard.test.tsx": 96,
     "apps/web/src/surfaces/review/ProposalReviewCard.tsx": 272,
     "apps/web/src/surfaces/review/ProposalReviewPanel.tsx": 184,
+    "apps/web/src/surfaces/studio/CodeSurface.test.tsx": 114,
+    "apps/web/src/surfaces/studio/CodeSurface.tsx": 630,
+    "apps/web/src/surfaces/studio/codeSurfaceApi.ts": 82,
+    "apps/web/src/surfaces/studio/codeTokens.ts": 177,
+    "apps/web/src/surfaces/studio/EditorCollections.test.tsx": 13,
+    "apps/web/src/surfaces/studio/editorJsonLiveDiff.ts": 92,
+    "apps/web/src/surfaces/studio/EditorPanel.test.tsx": 16,
+    "apps/web/src/surfaces/studio/EditorPanel.tsx": 644,
+    "apps/web/src/surfaces/studio/tsCompilerOptions.ts": 62,
     "apps/web/src/telemetry.test.ts": 201,
     "apps/web/src/telemetry.ts": 968,
     "apps/web/src/TelemetryCharts.test.tsx": 26,
@@ -595,7 +599,6 @@
     "apps/web/src/TelemetryOverview.tsx": 131,
     "apps/web/src/telemetryTrends.ts": 82,
     "apps/web/src/TelemetryTrendsPanel.tsx": 49,
-    "apps/web/src/tsCompilerOptions.ts": 62,
     "apps/web/src/UnpublishedPlayView.tsx": 126,
     "apps/web/src/useDeviceTilt.test.tsx": 328,
     "apps/web/src/useDeviceTilt.ts": 809,
