+++ eslint-rules/module-size-baseline.json
@@ -515,24 +515,13 @@
     "apps/web/src/ClaimHandleModal.tsx": 185,
     "apps/web/src/ClosedBetaSplash.test.ts": 431,
     "apps/web/src/ClosedBetaSplash.tsx": 181,
-    "apps/web/src/codeActionsMatch.ts": 31,
-    "apps/web/src/CodeActionsMenu.tsx": 433,
     "apps/web/src/codeMirrorColors.test.ts": 33,
-    "apps/web/src/codeMirrorColors.ts": 38,
     "apps/web/src/CodeMirrorEditor.test.tsx": 136,
-    "apps/web/src/CodeMirrorEditor.tsx": 993,
     "apps/web/src/codeMirrorSearchPanel.test.ts": 81,
-    "apps/web/src/codeMirrorSearchPanel.ts": 258,
     "apps/web/src/CodeSurface.actionsMenu.test.tsx": 294,
     "apps/web/src/CodeSurface.languageService.test.tsx": 395,
     "apps/web/src/CodeSurface.test.tsx": 1410,
-    "apps/web/src/CodeSurface.tsx": 1681,
-    "apps/web/src/codeSurfaceApi.ts": 285,
     "apps/web/src/codeSurfaceEditorState.test.ts": 33,
-    "apps/web/src/codeSurfaceEditorState.ts": 35,
-    "apps/web/src/codeSurfaceLanguageService.ts": 52,
-    "apps/web/src/codeSurfaceSessionState.ts": 32,
-    "apps/web/src/codeTokens.ts": 217,
     "apps/web/src/conceptLength.test.ts": 14,
     "apps/web/src/conceptLength.ts": 6,
     "apps/web/src/connectApi.ts": 175,
@@ -556,25 +545,20 @@
     "apps/web/src/CreatorQA.test.tsx": 577,
     "apps/web/src/CreatorQA.tsx": 598,
     "apps/web/src/CreatorStudioView.claimGate.test.tsx": 140,
-    "apps/web/src/CreatorStudioView.codeActions.test.tsx": 272,
+    "apps/web/src/CreatorStudioView.codeActions.test.tsx": 276,
     "apps/web/src/CreatorStudioView.handoff.test.tsx": 274,
     "apps/web/src/CreatorStudioView.test.tsx": 2143,
     "apps/web/src/CreatorStudioView.theater.test.tsx": 145,
     "apps/web/src/CreatorStudioView.tsx": 2241,
     "apps/web/src/diffLines.test.ts": 50,
-    "apps/web/src/diffLines.ts": 46,
     "apps/web/src/editorBridge.test.ts": 23,
     "apps/web/src/editorBridge.ts": 521,
     "apps/web/src/EditorCollections.test.tsx": 595,
     "apps/web/src/editorContentTools.ts": 445,
     "apps/web/src/editorJsonLiveDiff.test.ts": 127,
-    "apps/web/src/editorJsonLiveDiff.ts": 88,
     "apps/web/src/EditorPanel.test.tsx": 285,
-    "apps/web/src/EditorPanel.tsx": 1126,
     "apps/web/src/editorParamsScrub.test.ts": 115,
-    "apps/web/src/editorParamsScrub.ts": 78,
     "apps/web/src/EditorSurface.test.tsx": 74,
-    "apps/web/src/EditorSurface.tsx": 319,
     "apps/web/src/EditProfileModal.test.tsx": 284,
     "apps/web/src/EditProfileModal.tsx": 286,
     "apps/web/src/featuredApi.ts": 15,
@@ -656,7 +640,6 @@
     "apps/web/src/notificationsApi.test.ts": 52,
     "apps/web/src/notificationsApi.ts": 83,
     "apps/web/src/NumberScrubber.test.tsx": 98,
-    "apps/web/src/NumberScrubber.tsx": 86,
     "apps/web/src/oauthReturn.test.ts": 36,
     "apps/web/src/oauthReturn.ts": 25,
     "apps/web/src/pageTitle.test.ts": 129,
@@ -839,6 +822,62 @@
     "apps/web/src/surfaces/review/ReviewDesk.test.tsx": 279,
     "apps/web/src/surfaces/review/ReviewDesk.tsx": 612,
     "apps/web/src/surfaces/review/reviewTypes.ts": 32,
+    "apps/web/src/surfaces/studio/codeActionsMatch.ts": 31,
+    "apps/web/src/surfaces/studio/CodeActionsMenu.tsx": 433,
+    "apps/web/src/surfaces/studio/codeMirrorColors.test.ts": 33,
+    "apps/web/src/surfaces/studio/codeMirrorColors.ts": 38,
+    "apps/web/src/surfaces/studio/CodeMirrorEditor.test.tsx": 136,
+    "apps/web/src/surfaces/studio/CodeMirrorEditor.tsx": 993,
+    "apps/web/src/surfaces/studio/codeMirrorSearchPanel.test.ts": 81,
+    "apps/web/src/surfaces/studio/codeMirrorSearchPanel.ts": 258,
+    "apps/web/src/surfaces/studio/CodeSurface.actionsMenu.test.tsx": 294,
+    "apps/web/src/surfaces/studio/CodeSurface.explorer.test.tsx": 293,
+    "apps/web/src/surfaces/studio/CodeSurface.languageService.test.tsx": 395,
+    "apps/web/src/surfaces/studio/CodeSurface.test.tsx": 1410,
+    "apps/web/src/surfaces/studio/CodeSurface.tsx": 1602,
+    "apps/web/src/surfaces/studio/codeSurfaceApi.ts": 285,
+    "apps/web/src/surfaces/studio/codeSurfaceArchive.test.ts": 143,
+    "apps/web/src/surfaces/studio/codeSurfaceArchive.ts": 171,
+    "apps/web/src/surfaces/studio/codeSurfaceDrop.test.ts": 36,
+    "apps/web/src/surfaces/studio/codeSurfaceDrop.ts": 78,
+    "apps/web/src/surfaces/studio/codeSurfaceEditorState.test.ts": 33,
+    "apps/web/src/surfaces/studio/codeSurfaceEditorState.ts": 35,
+    "apps/web/src/surfaces/studio/CodeSurfaceExplorerTree.tsx": 303,
+    "apps/web/src/surfaces/studio/codeSurfaceLanguageBind.test.ts": 18,
+    "apps/web/src/surfaces/studio/codeSurfaceLanguageBind.ts": 47,
+    "apps/web/src/surfaces/studio/codeSurfaceLanguageService.ts": 48,
+    "apps/web/src/surfaces/studio/codeSurfacePaths.test.ts": 48,
+    "apps/web/src/surfaces/studio/codeSurfacePaths.ts": 128,
+    "apps/web/src/surfaces/studio/codeSurfaceSessionState.ts": 32,
+    "apps/web/src/surfaces/studio/codeSurfaceTreeApply.test.ts": 33,
+    "apps/web/src/surfaces/studio/codeSurfaceTreeApply.ts": 54,
+    "apps/web/src/surfaces/studio/CodeSurfaceTreeDialogs.tsx": 263,
+    "apps/web/src/surfaces/studio/codeSurfaceTreeModel.test.ts": 27,
+    "apps/web/src/surfaces/studio/codeSurfaceTreeModel.ts": 111,
+    "apps/web/src/surfaces/studio/codeSurfaceUpload.test.ts": 81,
+    "apps/web/src/surfaces/studio/codeSurfaceUpload.ts": 158,
+    "apps/web/src/surfaces/studio/codeTokens.ts": 217,
+    "apps/web/src/surfaces/studio/diffLines.test.ts": 50,
+    "apps/web/src/surfaces/studio/diffLines.ts": 46,
+    "apps/web/src/surfaces/studio/EditorCollections.test.tsx": 595,
+    "apps/web/src/surfaces/studio/editorJsonLiveDiff.test.ts": 127,
+    "apps/web/src/surfaces/studio/editorJsonLiveDiff.ts": 88,
+    "apps/web/src/surfaces/studio/EditorPanel.surfaceMode.test.ts": 45,
+    "apps/web/src/surfaces/studio/EditorPanel.test.tsx": 285,
+    "apps/web/src/surfaces/studio/EditorPanel.tsx": 1122,
+    "apps/web/src/surfaces/studio/editorParamsScrub.test.ts": 115,
+    "apps/web/src/surfaces/studio/editorParamsScrub.ts": 78,
+    "apps/web/src/surfaces/studio/EditorSurface.test.tsx": 74,
+    "apps/web/src/surfaces/studio/EditorSurface.tsx": 319,
+    "apps/web/src/surfaces/studio/editorSurfaceMode.ts": 16,
+    "apps/web/src/surfaces/studio/NumberScrubber.test.tsx": 98,
+    "apps/web/src/surfaces/studio/NumberScrubber.tsx": 86,
+    "apps/web/src/surfaces/studio/tsCompilerOptions.ts": 20,
+    "apps/web/src/surfaces/studio/tsWorker.ts": 37,
+    "apps/web/src/surfaces/studio/useCodeSurfaceTree.ts": 482,
+    "apps/web/src/surfaces/studio/useEditorDocument.ts": 161,
+    "apps/web/src/surfaces/studio/webmcp.test.ts": 305,
+    "apps/web/src/surfaces/studio/webmcp.ts": 329,
     "apps/web/src/telemetry.test.ts": 285,
     "apps/web/src/telemetry.ts": 273,
     "apps/web/src/TelemetryCharts.test.tsx": 197,
@@ -849,15 +888,12 @@
     "apps/web/src/telemetryTrends.ts": 181,
     "apps/web/src/TelemetryTrendsPanel.tsx": 277,
     "apps/web/src/TransparencySection.tsx": 63,
-    "apps/web/src/tsCompilerOptions.ts": 20,
-    "apps/web/src/tsWorker.ts": 38,
     "apps/web/src/UnpublishedPlayView.tsx": 103,
     "apps/web/src/useClampToViewport.test.ts": 67,
     "apps/web/src/useClampToViewport.ts": 27,
     "apps/web/src/useDeviceTilt.test.tsx": 316,
     "apps/web/src/useDeviceTilt.ts": 329,
     "apps/web/src/useDocumentTitle.ts": 12,
-    "apps/web/src/useEditorDocument.ts": 161,
     "apps/web/src/useInView.test.ts": 192,
     "apps/web/src/useInView.ts": 58,
     "apps/web/src/usePageScrolling.test.ts": 203,
@@ -884,7 +920,6 @@
     "apps/web/src/VoteWidget.test.tsx": 159,
     "apps/web/src/VoteWidget.tsx": 88,
     "apps/web/src/webmcp.test.ts": 305,
-    "apps/web/src/webmcp.ts": 329,
     "apps/web/src/welcomeProgress.test.ts": 40,
     "apps/web/src/welcomeProgress.ts": 42,
     "apps/web/src/world.test.tsx": 322,
